Bitboard::from()
